refactor(pedagogy): split dialogue_manager turn.rs into responsibility submodules
Behaviour-preserving split of the 746-line per-turn hot path into a
directory module, every file under 500 lines:

- turn/mod.rs (138) — module doc + the respond_to / respond_to_streaming
  orchestrator.
- turn/prompt.rs (132) — record_child_turn, build_turn_prompt,
  record_primer_turn. build_turn_prompt is exercised directly by
  dialogue_manager/tests/turn_tests.rs, so its visibility widens from
  pub(super) to pub(in crate::dialogue_manager) — same effective reach
  as before the move (the old pub(super) resolved to dialogue_manager).
- turn/stream.rs (137) — StreamOutcome, stream_inference_response
  (now file-private; its only caller moved with it), run_recovery_loop.
- turn/persist.rs (98) — persist_turn + spawn_embedding_task.
- turn/spawn_tasks.rs (284) — spawn_classification_task +
  spawn_post_response_task.

All method bodies and doc comments are verbatim relocations; only
imports, visibility qualifiers, and the previously parent-relative
super:: paths (now crate::dialogue_manager::…) changed. External pub
surface identical (respond_to, respond_to_streaming). Zero churn in the
test files.

Verification: cargo test -p primer-pedagogy 216 passed / 0 failed
(identical to pre-split baseline); workspace clippy -D warnings clean;
fmt clean; full workspace suite 51 x "test result: ok", 0 failures;
pub-surface diff empty.

CLAUDE.md: dialogue_manager module-layout note updated + two stale
turn.rs path references fixed (stream_inference_response,
effective_context_window_turns call site).
